# hostinger-deploy/docs/openapi.yaml
openapi: 3.0.3
info:
  title: QA Demo Store API
  version: "3.0.0"
  description: |
    REST API for the QA Demo Store (v3).

    **Base paths**
    - Production (PHP): `https://<your-domain>/api`
    - Preview (FastAPI mirror): `${REACT_APP_BACKEND_URL}/api`

    **Authentication**
    1. Call `POST /auth/login` with username and password.
    2. Use the returned token on protected routes: `Authorization: Bearer <token>`.
    3. Tokens are opaque (64 hex chars), valid for 24 hours.

    Legacy `.php` aliases (e.g. `/products.php`) are supported alongside clean paths.
servers:
  - url: https://seleniumbootcamp.in/api
    description: Production (Hostinger)
  - url: /api
    description: Same origin (relative)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Token from POST /auth/login (64 hex chars, 24h TTL)
  schemas:
    User:
      type: object
      required: [id, username, email, first_name, last_name, locked]
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: standard_user
        email:
          type: string
          format: email
          example: standard@demo.test
        first_name:
          type: string
          example: Standard
        last_name:
          type: string
          example: User
        locked:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: "2026-02-28T06:30:00+00:00"
    Product:
      type: object
      required: [id, name, description, price, image_url, category, stock]
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Sauce Labs Backpack
        description:
          type: string
        price:
          type: number
          format: float
          example: 29.99
        image_url:
          type: string
          format: uri
        category:
          type: string
          example: Bags
        stock:
          type: integer
          example: 25
...
    Order:
      type: object
      required: [id, user_id, full_name, address, city, state, pincode, phone, total, created_at, items]
      properties:
        id:
          type: integer
          example: 12
        user_id:
          type: integer
          example: 1
        full_name:
          type: string
          example: Alice Smith
        address:
          type: string
          example: 123 Main
        city:
          type: string
          example: Mumbai
        state:
          type: string
          example: MH
        pincode:
          type: string
          example: "400001"
        phone:
          type: string
          example: "+91 9876543210"
        total:
          type: number
          format: float
          example: 69.97
        created_at:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: standard_user
        password:
          type: string
          format: password
          example: secret_sauce
    LoginResponse:
      type: object
      required: [token, expires_at, user]
      properties:
        token:
          type: string
          description: 64-character hex token
          example: 8b3f0123456789abcdef0123456789abcdef0123456789abcdef0123456789ab
        expires_at:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
    SignupRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 60
          pattern: "^[A-Za-z0-9_.-]+$"
          example: alice
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          format: password
          minLength: 6
          example: hunter22
    SignupResponse:
      type: object
      required: [success, message, user]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Account created successfully
        user:
          $ref: "#/components/schemas/User"
    OrderItemIn:
      type: object
      required: [product_id, quantity]
      properties:
        product_id:
          type: integer
          example: 1
        quantity:
          type: integer
          minimum: 1
          example: 2
    CreateOrderRequest:
      type: object
      required: [full_name, address, city, state, pincode, phone, items]
      properties:
        full_name:
          type: string
          minLength: 2
          maxLength: 120
          example: Alice Smith
        address:
          type: string
          minLength: 2
          maxLength: 255
          example: 123 Main
        city:
          type: string
          minLength: 1
          maxLength: 80
          example: Mumbai
        state:
          type: string
          minLength: 1
          maxLength: 80
          example: Maharashtra
        pincode:
          type: string
          minLength: 3
          maxLength: 20
          example: "400001"
        phone:
          type: string
          pattern: "^[+\\d][\\d\\s\\-()]{5,19}$"
          example: "+91 9876543210"
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItemIn"
    CreateOrderResponse:
      type: object
      required: [success, order_id, total]
      properties:
        success:
          type: boolean
          example: true
        order_id:
          type: integer
          example: 13
        total:
          type: number
          format: float
          example: 61.97
    ProductsListResponse:
      type: object
      required: [products]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
    ProductResponse:
      type: object
      required: [product]
      properties:
        product:
          $ref: "#/components/schemas/Product"
    ProfileResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    OrdersListResponse:
      type: object
      required: [orders]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: Missing bearer token
paths:
...
  /products:
    get:
      tags: [Public]
      summary: List all products or get one by id
      operationId: listOrGetProducts
      description: Omit `id` to list the catalog. Pass `id` to fetch a single product.
      parameters:
        - name: id
          in: query
          required: false
          schema:
            type: integer
          description: Product primary key
          example: 1
      responses:
        "200":
          description: Product list or single product
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ProductsListResponse"
                  - $ref: "#/components/schemas/ProductResponse"
              examples:
                list:
                  summary: All products
                  value:
                    products:
                      - id: 1
                        name: Sauce Labs Backpack
                        description: Carry.allTheThings() with this sleek backpack.
                        price: 29.99
                        image_url: https://images.unsplash.com/photo-1553062407-98eeb64c6a62?w=600&q=80
                        category: Bags
                        stock: 25
                single:
                  summary: One product
                  value:
                    product:
                      id: 1
                      name: Sauce Labs Backpack
                      description: Carry.allTheThings() with this sleek backpack.
                      price: 29.99
                      image_url: https://images.unsplash.com/photo-1553062407-98eeb64c6a62?w=600&q=80
                      category: Bags
                      stock: 25
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products.php:
    get:
      tags: [Public]
      summary: List all products or get one by id (.php alias)
      operationId: listOrGetProductsPhp
      parameters:
        - name: id
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Product list or single product
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ProductsListResponse"
                  - $ref: "#/components/schemas/ProductResponse"
        "404":
          description: Product not found
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      operationId: login
      description: Exchange username and password for a Bearer token (24-hour TTL).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: Username and password do not match any user in this service
        "403":
          description: Account locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: Sorry, this user has been locked out.
  /auth/login.php:
    post:
      tags: [Auth]
      summary: Login (.php alias)
      operationId: loginPhp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
        "403":
          description: Account locked
  /signup:
    post:
      tags: [Auth]
      summary: Sign up
      operationId: signup
      description: Create a new user account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignupResponse"
        "400":
          description: Validation error or duplicate username/email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /signup.php:
    post:
      tags: [Auth]
      summary: Sign up (.php alias)
      operationId: signupPhp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignupResponse"
        "400":
          description: Validation error or duplicate username/email
  /profile:
    get:
      tags: [Protected]
      summary: Current user profile
      operationId: getProfile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "401":
          description: Missing, invalid, or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /profile.php:
    get:
      tags: [Protected]
      summary: Current user profile (.php alias)
      operationId: getProfilePhp
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "401":
          description: Unauthorized
  /orders:
    get:
      tags: [Protected]
      summary: List orders
      operationId: listOrders
      description: Returns orders for the authenticated user. Pass `user_id` to filter by another user.
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: query
          required: false
          schema:
            type: integer
          description: Filter by user id; defaults to caller
          example: 1
      responses:
        "200":
          description: Order list (newest first)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersListResponse"
        "401":
          description: Missing, invalid, or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags: [Protected]
      summary: Create order
      operationId: createOrder
      description: Order total is always recomputed server-side from current product prices.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
            example:
              full_name: Alice Smith
              address: 123 Main
              city: Mumbai
              state: Maharashtra
              pincode: "400001"
              phone: "+91 9876543210"
              items:
                - product_id: 1
                  quantity: 1
                - product_id: 3
                  quantity: 2
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateOrderResponse"
        "400":
          description: Validation error, empty cart, or unknown product id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing, invalid, or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders.php:
    get:
      tags: [Protected]
      summary: List orders (.php alias)
      operationId: listOrdersPhp
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersListResponse"
        "401":
          description: Unauthorized

    post:
      tags: [Protected]
      summary: Create order (.php alias)
      operationId: createOrderPhp
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateOrderResponse"
        "400":
          description: Bad request
        "401":
          description: Unauthorized